refactor: NewCardScreen 구조 리팩토링 보충
