+++ README.md
@@ -7,6 +7,11 @@
 - 카드 데이터 저장소 구현
 - NewCardScreen 구조 리팩토링
 - CardListScreen 껍데기 구현
+- 카드 추가 버튼 UI 구현
+- 카드 내용 UI 구현
+- 카드가 없는 케이스 화면 구현
+- 카드가 한 개인 케이스 화면 구현
+- 카드가 n 개인 케이스 화면 구혐
 - 시작 화면을 카드 CardListScreen으로 변경
 - CardListScreen UI 및 ViewModel 구현
 - CardListScreen 기능 구현
